requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
django==4.2.7; python_version >= '3.8'
djangorestframework==3.14.0; python_version >= '3.6'
psycopg==3.1.13; python_version >= '3.7'
pytz==2023.3.post1
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.8.0; python_version >= '3.8'
